+++ source/displayfile.py
@@ -4,7 +4,12 @@
 Módulo para o handler do display file.
 '''
 
-from source.wireframe import Object, Point, Line, Wireframe
+import gi
+
+from source.wireframe import Object
+
+gi.require_version("Gtk", "3.0")
+from gi.repository import Gtk
 
 
 class DisplayFileHandler():
@@ -16,9 +21,13 @@ class DisplayFileHandler():
     # Atributos públicos
     objects: list
 
-    def __init__(self) -> None:
+    # Atributos privados
+    _display_file_list: Gtk.ListStore
+
+    def __init__(self, display_file_list) -> None:
 
         self.objects = []
+        self._display_file_list = display_file_list
 
     # Métodos utilitários
     def add_object(self, obj: Object) -> None:
@@ -27,6 +36,7 @@ class DisplayFileHandler():
         '''
 
         self.objects.append(obj)
+        self._display_file_list.append([obj.name, "()"])
 
     # Por enquanto o id é o nome
     def remove_object(self, identification: str) -> None:
@@ -39,3 +49,11 @@ class DisplayFileHandler():
             if obj.identification == identification:
                 self.objects.remove(obj)
                 break
+
+    def remove_all(self) -> None:
+        '''
+        Remove todos os objetos.
+        '''
+
+        self.objects.clear()
+        self._display_file_list.clear()